Profil info, Auth update, wishlist update, password changing

+++ night-bazaar-app/lib/view/product_detail.dart
@@ -1,3 +1,4 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get/get.dart';
@@ -37,7 +38,6 @@ class ProductDetail extends ConsumerStatefulWidget {
 class _ProductDetailState extends ConsumerState<ProductDetail> {
   WishlistService wishlistService = WishlistService();
   bool isSaved = false; // Yeni bool değişkeni tanımlanıyor
-  
 
   Future<void> checkIfProductIsSaved() async {
     final userUid = wishlistService.getCurrentUserUid();
@@ -132,28 +132,51 @@ class _ProductDetailState extends ConsumerState<ProductDetail> {
               ),
             ),
             GestureDetector(
-              onTap: () {
-                if (isSaved) {
-                  setState(() {
-                    isSaved = !isSaved;
-                  });
-                  wishlistService.removeFromWishlist(widget.productId);
+              onTap: () async {
+                final user = FirebaseAuth.instance.currentUser;
+                if (user != null) {
+                  // Kullanıcı oturum açmış, favorilere ekleme veya kaldırma işlemini yapabilirsiniz
+                  if (isSaved) {
+                    setState(() {
+                      isSaved = !isSaved;
+                    });
+                    wishlistService.removeFromWishlist(widget.productId);
+                  } else {
+                    setState(() {
+                      isSaved = !isSaved;
+                    });
+                    // Ürün verilerini kullanarak Product örneği oluşturun
+                    Product productToAdd = Product(
+                      productId: widget.productId,
+                      image: widget.image,
+                      title: widget.title,
+                      price: widget.price,
+                      star: widget.star,
+                      isSaved: isSaved,
+                      descTitle: widget.descTitle,
+                      desc: widget.desc,
+                    );
+                    wishlistService.addToWishlist(productToAdd);
+                  }
                 } else {
-                  setState(() {
-                    isSaved = !isSaved;
-                  });
-                  // Ürün verilerini kullanarak Product örneği oluşturun
-                  Product productToAdd = Product(
-                    productId: widget.productId,
-                    image: widget.image,
-                    title: widget.title,
-                    price: widget.price,
-                    star: widget.star,
-                    isSaved: isSaved,
-                    descTitle: widget.descTitle,
-                    desc: widget.desc,
-                  );
-                  wishlistService.addToWishlist(productToAdd);
+                  // Kullanıcı oturum açmamış, uyarı mesajı gösterin
+                  showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return AlertDialog(
+                          title: Text("Oturum Açınız"),
+                          content: Text(
+                              "Favorilere eklemek için lütfen oturum açınız."),
+                          actions: [
+                            TextButton(
+                              onPressed: () {
+                                Navigator.of(context).pop();
+                              },
+                              child: Text("Kapat"),
+                            ),
+                          ],
+                        );
+                      });
                 }
               },
               child: isSaved
